requirements.txt
python-dotenv
requests==2.32.3
ytmusicapi==1.8.2
